updates the directory based on review
