+++ src/crypto/utils/aes.ts
@@ -1,19 +1,29 @@
 const crypto = require('crypto');
 
-//Encryption
-export const _encryptData = function (data: Object, cipherKey: Object): Promise<Buffer> {
-    
+export const _generateCipherKey = () => {
+  try {
+    const seed = crypto.randomBytes(32).toString();
+    return new Promise((resolve) => {
+      const cipherKey = crypto.createHash('sha256').update(seed).digest();
+      resolve(cipherKey);
+    });
+  } catch (err) {
+    console.error('Error while generating symmetric key:', err);
+    return null;
+  }
+};
+
+export const _aesEncryption = async (data: Object, cipherKey: Object): Promise<Buffer> => {
   return new Promise((resolve) => {
     let iv = crypto.randomBytes(16);
     const cipher = crypto.createCipheriv('aes256', cipherKey, iv);
     const encryptedData: Buffer = Buffer.concat([iv, cipher.update(data), cipher.final()]);
     resolve(encryptedData);
   });
+}
 
-};
 
-//Decryption
-export const _decryptData = async function (data: any, cipherKey: any): Promise<Object> {
+export const _aesDecryption = async (data: any, cipherKey: any): Promise<Object>  => {
 
   let encryptedData = Buffer.from(data, 'hex')
   const iv = encryptedData.slice(0, 16);
@@ -24,4 +34,4 @@ export const _decryptData = async function (data: any, cipherKey: any): Promise<
     resolve(decryptedData);
   });
 
-};
+}
